feat: reverse weight

+++ contract/Schrodinger/SchrodingerContractConstants.cs
@@ -28,7 +28,7 @@ public static class SchrodingerContractConstants
     public const long DefaultImageMaxCount = 10;
     public const long DefaultTraitValueMaxCount = 100;
     public const long DefaultAttributeMaxLength = 80;
-    public const int DefaultMaxAttributeWeight = 1000;
+    public const long DefaultMaxAttributeWeight = 1000;
     public const int DefaultMaxAttributePerGen = 1;
     public const int DefaultMaxAttributeTraitTypeCount = 50;
